+++ src/games.c
@@ -56,14 +56,16 @@ static char* stringlist053[] = { "openarena", NULL };
 static char* stringlist054[] = { "openttd", NULL };
 static char* stringlist055[] = { "baseq3", NULL };
 static char* stringlist056[] = { "q3rally", NULL };
-static char* stringlist057[] = { "q3ut4", NULL };
-static char* stringlist058[] = { "ioUrbanTerror", "Quake3-UrT", "urbanterror", NULL };
-static char* stringlist059[] = { "rq3", NULL };
-static char* stringlist060[] = { "Reaction", NULL };
-static char* stringlist061[] = { "smokinguns", NULL };
-static char* stringlist062[] = { "smokinguns", NULL };
-static char* stringlist063[] = { "arena", NULL };
-static char* stringlist064[] = { "alienarena", NULL };
+static char* stringlist057[] = { "wop", NULL };
+static char* stringlist058[] = { "worldofpadman", NULL };
+static char* stringlist059[] = { "q3ut4", NULL };
+static char* stringlist060[] = { "ioUrbanTerror", "Quake3-UrT", "urbanterror", NULL };
+static char* stringlist061[] = { "rq3", NULL };
+static char* stringlist062[] = { "Reaction", NULL };
+static char* stringlist063[] = { "smokinguns", NULL };
+static char* stringlist064[] = { "smokinguns", NULL };
+static char* stringlist065[] = { "arena", NULL };
+static char* stringlist066[] = { "alienarena", NULL };
 struct game games[] = {
   {
     type                : Q1_SERVER,
@@ -1133,6 +1135,38 @@ struct game games[] = {
     command             : stringlist056,
   },
   {
+    type                : WOP_SERVER,
+    flags               : GAME_CONNECT | GAME_PASSWORD | GAME_RCON | GAME_QUAKE3_MASTERPROTOCOL | GAME_Q3COLORS,
+    name                : "World of Padman",
+    default_port        : 27960,
+    default_master_port : 27950,
+    id                  : "WOPS",
+    qstat_str           : "WOPS",
+    qstat_option        : "-wops",
+    qstat_master_option : "-wopm",
+    icon                : "wop.xpm",
+    parse_player        : q3_parse_player,
+    parse_server        : quake_parse_server,
+    analyze_serverinfo  : q3_analyze_serverinfo,
+    config_is_valid     : config_is_valid_generic,
+    exec_client         : q3_exec,
+    custom_cfgs         : quake_custom_cfgs,
+    save_info           : quake_save_info,
+    init_maps           : q3_init_maps,
+    has_map             : quake_has_map,
+    get_mapshot         : q3_get_mapshot,
+    arch_identifier     : "version",
+    identify_cpu        : identify_cpu,
+    identify_os         : identify_os,
+    cmd_or_dir_changed  : tremulous_cmd_or_dir_changed,
+    prefs_load          : q3_prefs_load_common,
+    update_prefs        : tremulous_update_prefs,
+    default_home        : "~/.WoPadman",
+    pd                  : &wop_private,
+    main_mod            : stringlist057,
+    command             : stringlist058,
+  },
+  {
     type                : IOURT_SERVER,
     flags               : GAME_CONNECT | GAME_PASSWORD | GAME_RCON | GAME_QUAKE3_MASTERPROTOCOL | GAME_Q3COLORS,
     name                : "UrbanTerror",
@@ -1161,8 +1195,8 @@ struct game games[] = {
     update_prefs        : tremulous_update_prefs,
     default_home        : "~/.q3a",
     pd                  : &iourt_private,
-    main_mod            : stringlist057,
-    command             : stringlist058,
+    main_mod            : stringlist059,
+    command             : stringlist060,
   },
   {
     type                : REACTION_SERVER,
@@ -1193,8 +1227,8 @@ struct game games[] = {
     update_prefs        : tremulous_update_prefs,
     default_home        : "~/.Reaction",
     pd                  : &reaction_private,
-    main_mod            : stringlist059,
-    command             : stringlist060,
+    main_mod            : stringlist061,
+    command             : stringlist062,
   },
   {
     type                : SMOKINGUNS_SERVER,
@@ -1225,8 +1259,8 @@ struct game games[] = {
     update_prefs        : tremulous_update_prefs,
     default_home        : "~/.smokinguns",
     pd                  : &smokinguns_private,
-    main_mod            : stringlist061,
-    command             : stringlist062,
+    main_mod            : stringlist063,
+    command             : stringlist064,
   },
   {
     type                : GPS_SERVER,
@@ -1273,8 +1307,8 @@ struct game games[] = {
     update_prefs        : tremulous_update_prefs,
     default_home        : "~/.codered/",
     pd                  : &alienarena_private,
-    main_mod            : stringlist063,
-    command             : stringlist064,
+    main_mod            : stringlist065,
+    command             : stringlist066,
   },
   {
     type                : UNKNOWN_SERVER,