+++ src/pref.c
@@ -530,6 +530,21 @@ static struct q3_common_prefs_s q3rally_prefs =
   defproto : "71",
 };
 
+static const char* wop_masterprotocols[] = {
+        "auto",
+        "71",
+        "70",
+        "69",
+        "68",
+        NULL
+};
+
+static struct q3_common_prefs_s wop_prefs =
+{
+  protocols: wop_masterprotocols,
+  defproto : "71",
+};
+
 static const char* iourt_masterprotocols[] = {
         "auto",
         "68 - v1.35",
@@ -3238,6 +3253,7 @@ static struct q3_common_prefs_s* get_pref_widgets_for_game(enum server_type type
     case UNVANQUISHED_SERVER: return &unvanquished_prefs;
     case OPENARENA_SERVER: return &openarena_prefs;
     case Q3RALLY_SERVER: return &q3rally_prefs; 
+    case WOP_SERVER: return &wop_prefs; 
     case IOURT_SERVER: return &iourt_prefs; 
     case REACTION_SERVER: return &reaction_prefs; 
     case SMOKINGUNS_SERVER: return &smokinguns_prefs; 
@@ -4810,6 +4826,7 @@ static struct generic_prefs* new_generic_prefs (void) {
   new_genprefs[UNVANQUISHED_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
   new_genprefs[OPENARENA_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
   new_genprefs[Q3RALLY_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
+  new_genprefs[WOP_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
   new_genprefs[IOURT_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
   new_genprefs[REACTION_SERVER].add_options_to_notebook = add_q3_options_to_notebook;
   new_genprefs[SMOKINGUNS_SERVER].add_options_to_notebook = add_q3_options_to_notebook;